THE MESSAGE

+++ db/schema.rb
@@ -17,8 +17,8 @@ ActiveRecord::Schema.define(version: 2021_02_19_155458) do
 
   create_table "bookmarks", force: :cascade do |t|
     t.text "comment"
-    t.bigint "movie_id", null: false
-    t.bigint "list_id", null: false
+    t.integer "movie_id", null: false
+    t.integer "list_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["list_id"], name: "index_bookmarks_on_list_id"
@@ -41,7 +41,7 @@ ActiveRecord::Schema.define(version: 2021_02_19_155458) do
   end
 
   create_table "reviews", force: :cascade do |t|
-    t.bigint "list_id", null: false
+    t.integer "list_id", null: false
     t.text "comment"
     t.integer "rating"
     t.datetime "created_at", precision: 6, null: false
